Correct the logic in move_dir()

+++ src/wp-admin/includes/file.php
@@ -1953,38 +1953,33 @@ function copy_dir( $from, $to, $skip_list = array() ) {
  *
  * @global WP_Filesystem_Base $wp_filesystem WordPress filesystem subclass.
  *
- * @param string $from        Source directory.
- * @param string $to          Destination directory.
- * @param string $working_dir Optional. Remote file source directory.
- *                            Default empty string.
+ * @param string $from Source directory.
+ * @param string $to   Destination directory.
  * @return true|WP_Error True on success, WP_Error on failure.
  */
-function move_dir( $from, $to, $working_dir = '' ) {
+function move_dir( $from, $to ) {
 	global $wp_filesystem;
 
-	if ( 'direct' === $wp_filesystem->method ) {
+	$result = false;
+
+	if ( 'direct' === $wp_filesystem->method && ! $wp_filesystem->is_virtual_box() ) {
 		$wp_filesystem->rmdir( $to );
-		if ( @rename( $from, $to ) ) {
-			return true;
-		}
+		$result = @rename( $from, $to );
 	}
 
-	if ( ! $wp_filesystem->is_dir( $to ) ) {
-		if ( ! $wp_filesystem->mkdir( $to, FS_CHMOD_DIR ) ) {
-
-			// Clear the working directory?
-			if ( ! empty( $working_dir ) ) {
-				$wp_filesystem->delete( $working_dir, true );
+	if ( ! $result ) {
+		if ( ! $wp_filesystem->is_dir( $to ) ) {
+			if ( ! $wp_filesystem->mkdir( $to, FS_CHMOD_DIR ) ) {
+				return new WP_Error( 'mkdir_failed_move_dir', __( 'Could not create directory.' ), $to );
 			}
-
-			return new WP_Error( 'mkdir_failed_move_dir', __( 'Could not create directory.' ), $to );
 		}
-	}
-	$result = copy_dir( $from, $to );
 
-	// Clear the working directory?
-	if ( ! empty( $working_dir ) ) {
-		$wp_filesystem->delete( $working_dir, true );
+		$result = copy_dir( $from, $to );
+
+		if ( ! is_wp_error( $result ) ) {
+			// Clear the source directory.
+			$wp_filesystem->delete( $from, true );
+		}
 	}
 
 	return $result;
